Fixed PyTest Fail
The test

TestRouteLayer::test_query_with_no_index

was failing because when it ran

assert route_layer("Anything").name is None

There were no Routes defined yet, and this meant that

route = [route for route in self.routes if route.name == top_class][0] was raising an error.

If no routes are defined then we can just defualt to RouteChoice().