Constructing List of New Route Names in Index Classes
We want to be able to retrieve a list of Routes from each Index class. For LocalIndex this is will be done by maintaining a list of Route objects and for PineconeIndex it will be done by checking against what's in the DB (we can't have a mantained local list as multiple users might be adding routes to the remote vector db).

Since, for LocalIndex, we need maintain a list of Routes, we also need to pass in the Route object itself to the add() method, not just a list of route names.  Since we're sending the Route object itself for LocalIndex.add(), we should also do it for PineconeIndex.add(), so that the self.index.add() methods can be called with consistent arguments in RouteLayer.

Subsequently, changes are made in LocalIndex and PineconeIndex to construct the list of Route names within those classes.